requirements.txt
# Core Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Web Framework (for backend server)
Flask==2.3.3
Flask-CORS==4.0.0

# Utilities
scikit-learn==1.3.0